Now Window is one class
